requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==3.2.6
idna==2.10
mysql-connector-python==8.0.29
mysqlclient==2.1.0
paramiko==2.10.3
protobuf==3.19.4
pycparser==2.21
PyNaCl==1.5.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sentry-sdk==1.5.10
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
